add a build cache example to the README
With a test-cache job on the repo's workflow, to show how it works.
We add a third party module dependency just to show what modules are
downloaded at each run.

The GOPATH example in the README was also slightly wrong, as it did not
set up the working directory inside GOPATH for the future steps. Fix
that.

Remove test-gopath from this repository's workflow, as having to
download dependencies via 'go get' would make it non-reproducible and
perhaps fail in the future. The example in the README is enough moving
forward, as everyone should be using modules by now.

Fixes #7.