fix caching for windows and tweak cache keys
It turns out that the globbing used by actions/cache does not actually
support reading env vars, and so the caching of GOCACHE on windows never
worked. Use the full path relative to $HOME.

We also start suggesting the inclusion of the Go version as part of the
cache key, because GOCACHE is different per go version.

Finally, bump the versions of setup-go and checkout.